pagination design and functionality issue fixed with old component
