+++ src/components/Estimates_dep/existingTable.jsx
@@ -28,7 +28,7 @@ import { useNavigate } from "react-router-dom";
 import { calculateAreaAndPerimeter, calculateTotal } from "../../utilities/common";
 import { DataGrid } from "@mui/x-data-grid";
 import { EstimatesColumns } from "../../utilities/DataGridColumns";
-import Pagination from "../Pagination/index_dep";
+import Pagination from "../Pagination";
 import DeleteModal from "../Modal/deleteModal";
 import { getEstimatesListRefetch } from "../../redux/refetch";
 import { generateObjectForPDFPreview, renderMeasurementSides } from "../../utilities/estimates";
