+++ src/components/Pagination/index.js
@@ -1,87 +1,63 @@
-import { MAX_PAGES_DISPLAYED, itemsPerPage } from "@/utilities/constants";
 import { ArrowBack, ArrowForward } from "@mui/icons-material";
-import { Box, Button, TextField } from "@mui/material";
+import { Box, Button, useMediaQuery } from "@mui/material";
 
-const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isShowInput ,setIsShowInput }) => {
-    // const [page, setPage] = useState(1);
-    // const [inputPage, setInputPage] = useState("");
-    // const [isShowInput, setIsShowInput] = useState(false);
-  
-    const totalPages = Math.ceil(totalRecords / itemsPerPage);
-  
-    const getPageNumbersToShow = () => {
-      if (totalPages <= MAX_PAGES_DISPLAYED) {
-        return Array.from({ length: totalPages }, (_, index) => index + 1);
+const Pagination = ({ totalRecords, itemsPerPage, page, setPage }) => {
+  const isMobile =  useMediaQuery('(max-width:600px)');
+  const totalPages = Math.ceil(totalRecords / itemsPerPage);
+  const MAX_PAGES_DISPLAYED = isMobile ? 3 : 5; // Display fewer pages on mobile
+  const placeHolder = '...';
+  const getPageNumbersToShow = () => {
+    if (totalPages <= MAX_PAGES_DISPLAYED) {
+      return Array.from({ length: totalPages }, (_, index) => index + 1);
+    }
+
+    const pagesToShow = [];
+    const startPage = Math.max(1, page - 2); // Display three on the first side
+    const endPage = Math.min(totalPages, startPage + MAX_PAGES_DISPLAYED - 1);
+
+    if (startPage > 1) {
+      pagesToShow.push(1);
+      if (startPage > 2) {
+        pagesToShow.push(placeHolder); // Display ellipsis if there are skipped pages
       }
-  
-      const pagesToShow = [];
-      const startPage = Math.max(1, page - 2); // Display three on the first side
-      const endPage = Math.min(totalPages, startPage + MAX_PAGES_DISPLAYED - 1);
-  
-      if (startPage > 1) {
-        pagesToShow.push(1);
-        if (startPage > 2) {
-          pagesToShow.push("..."); // Display ellipsis if there are skipped pages
-        }
+    }
+
+    for (let i = startPage; i <= endPage; i++) {
+      pagesToShow.push(i);
+    }
+
+    if (endPage < totalPages) {
+      if (endPage < totalPages - 1) {
+        pagesToShow.push(placeHolder); // Display ellipsis if there are skipped pages
       }
-  
-      for (let i = startPage; i <= endPage; i++) {
-        pagesToShow.push(i);
-      }
-  
-      if (endPage < totalPages) {
-        if (endPage < totalPages - 1) {
-          pagesToShow.push("..."); // Display ellipsis if there are skipped pages
-        }
-        pagesToShow.push(totalPages);
-      }
-  
-      return pagesToShow;
-    };
-  
-    const pageNumbersToShow = getPageNumbersToShow();
-  
-    const handlePreviousPage = () => {
-      setPage((prevPage) => Math.max(prevPage - 1, 1));
-      setIsShowInput(false);
-    };
-  
-    const handleNextPage = () => {
-      setPage((prevPage) => Math.min(prevPage + 1, totalPages));
-      setIsShowInput(false);
-    };
-  
-    const handleInputChange = (event) => {
-      setInputPage(event.target.value);
-    };
-    const HandleShowInput = () => {
-      setIsShowInput(true);
-    };
-    const HandleShowRemoveInput = () => {
-      setIsShowInput(false);
-    };
-    const handleInputKeyPress = (event) => {
-      if (event.key === "Enter") {
-        const pageNumber = parseInt(inputPage, 10);
-        if (!isNaN(pageNumber) && pageNumber >= 1 && pageNumber <= totalPages) {
-          setPage(pageNumber);
-          setInputPage("");
-          setIsShowInput(false);
-        }
-      }
-    };
-  
-    const handleInputBlur = () => {
-      setInputPage("");
-    };
+      pagesToShow.push(totalPages);
+    }
+
+    return pagesToShow;
+  };
+
+  const pageNumbersToShow = getPageNumbersToShow();
+
+  const handlePreviousPage = () => {
+    setPage((prevPage) => Math.max(prevPage - 1, 1));
+  };
+
+  const handleNextPage = () => {
+    setPage((prevPage) => Math.min(prevPage + 1, totalPages));
+  };
+
+  const handleSetPageNumber = (pageNumber) => {
+   if(pageNumber !== placeHolder){
+    setPage(pageNumber);
+   }
+  }
   return (
     <Box
       sx={{
         display: "flex",
         justifyContent: "space-between",
-        padding: "10px",
+        padding: isMobile ? "5px" : "10px",
         borderTop: "1px solid #EAECF0",
-        width: "96%",
       }}
     >
       <Button
@@ -91,6 +67,9 @@ const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isS
           borderRadius: "8px",
           textTransform: "capitalize",
           fontWeight: 500,
+          background:'white',
+          minWidth: isMobile ? '0px' : 'auto',
+          padding: isMobile ? '6px' : '5px 15px',
           ":hover": {
             border: "1px solid #D0D5DD",
             color: "#344054",
@@ -101,20 +80,16 @@ const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isS
         disabled={page === 1}
       >
         <ArrowBack sx={{ color: "#344054", fontSize: 20, mr: 1 }} />
-        Previous
+       {isMobile ? '' : 'Previous'}
       </Button>
-      <Box sx={{ display: "flex", gap: 2 }}>
+      <Box sx={{ display: "flex", gap: isMobile ? "5px" : 2, }}>
         {pageNumbersToShow.map((pagenumber, index) => (
           <Box
             key={index}
-            onClick={() => {
-              if (typeof pagenumber === "number") {
-                setPage(pagenumber);
-                HandleShowRemoveInput();
-              }
-            }}
+            onClick={() => handleSetPageNumber(pagenumber)}
             sx={{
-              backgroundColor: page === pagenumber ? "#8477DA" : "white",
+              backgroundColor:
+              page === pagenumber ? "#8477DA" : "white",
               color: page === pagenumber ? "white" : "#8477DA",
               width: "40px",
               height: "40px",
@@ -122,34 +97,11 @@ const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isS
               display: "flex",
               alignItems: "center",
               justifyContent: "center",
-              border: "1px solid #8477DA",
-              cursor: typeof pagenumber === "number" ? "pointer" : "default",
+              cursor: pagenumber !== placeHolder ? "pointer" : 'text',
+              border: page === pagenumber ? "1px solid #8477DA" : "1px solid #D0D5DD"
             }}
           >
-            {isShowInput && pagenumber === "..." ? (
-              <TextField
-                size="small"
-                variant="outlined"
-                type="text"
-                value={inputPage}
-                onChange={handleInputChange}
-                onKeyPress={handleInputKeyPress}
-                onBlur={handleInputBlur}
-                inputProps={{
-                  min: 1,
-                  max: totalPages,
-                  inputMode: "numeric",
-                  pattern: "[0-9]*",
-                }}
-                sx={{ width: 60 }}
-              />
-            ) : pagenumber === "..." ? (
-              <div onClick={HandleShowInput} style={{ cursor: "pointer" }}>
-                {pagenumber}{" "}
-              </div>
-            ) : (
-              pagenumber
-            )}
+            {pagenumber}
           </Box>
         ))}
       </Box>
@@ -160,19 +112,22 @@ const NewPagination = ({ totalRecords, page, setPage,inputPage,setInputPage, isS
           borderRadius: "8px",
           textTransform: "capitalize",
           fontWeight: 500,
+          background:'white',
+          minWidth: isMobile ? '0px' : 'auto',
+          padding: isMobile ? '6px' : '5px 15px',
           ":hover": {
             border: "1px solid #D0D5DD",
             color: "#344054",
           },
         }}
         onClick={handleNextPage}
-        disabled={page === totalPages}
+        disabled={pageNumbersToShow?.length > page ? false : true}
       >
-        Next
+        {isMobile ? '' : 'Next'}
         <ArrowForward sx={{ color: "#344054", fontSize: 20, ml: 1 }} />
       </Button>
     </Box>
   );
 };
 
-export default NewPagination;
+export default Pagination;
