+++ src/utilities/common.js
@@ -910,3 +910,13 @@ export const getWeightByThickness = (variant, glassThickness, sqft) => {
       return {};
   }
 };
+
+export const debounce = (func, delay) => {
+  let timeout;
+  return (...args) => {
+    clearTimeout(timeout);
+    timeout = setTimeout(() => {
+      func(...args);
+    }, delay);
+  };
+};
